refactor(indexer/handlers): extract fpmm/liquidity.ts (Mint + Burn) (PR-S2)
Second slice of the handlers/fpmm.ts split (after PR-S1 #297). Move the
FPMM.Mint and FPMM.Burn handlers into the new
indexer-envio/src/handlers/fpmm/liquidity.ts (~101 lines).

EventHandlers.ts is updated to import the new module so Envio sees the
side-effect handler registrations at module load.

handlers/fpmm.ts: 773 → 678 lines.

Pure extraction. 420 tests pass.
