+++ indexer-envio/src/EventHandlers.ts
@@ -19,6 +19,7 @@ runStartupChecks();
 // Handler registrations (side-effect imports)
 import "./handlers/fpmm";
 import "./handlers/fpmm/factory";
+import "./handlers/fpmm/liquidity";
 import "./handlers/sortedOracles";
 import "./handlers/virtualPool";
 import "./handlers/feeToken";
